--- DCCA/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ['DCCA', 'NP_H']
