--- DPXA/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ['DPXA', 'P_H']
