+++ method/NP_H.py
