+++ method/P_H.py
