+++ switch.py
@@ -1,0 +1,6 @@
+import pandas as pd
+import numpy as np
+
+from method.DPXA import MF_DPXA
+from method.DCCA import MF_DCCA
+
